backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
Flask-SQLAlchemy==3.1.1
Pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
web3==6.20.1
